+++ Sample/InventoryPanelSample/Scripts/Inventory/IInventoryService.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+
+namespace KK.UI.UMG.Samples.Inventory
+{
+    public interface IInventoryService
+    {
+        event Action InventoryChanged;
+
+        IReadOnlyList<InventoryItemModel> GetItems();
+
+        bool TryUseItem(string itemId);
+    }
+}